Added debug print macro header.
Added a macro, that after the error is found prints the error message (string)
along with the number of the error found line and the file name.
Added the appropriate include in the common.c and changed the printfs to the debug print calls.
Changed the function call in main function from dci_lenghtrivviabandwidth to dci0_lenghtrivviabandwidth.